Minor bug fixes.

+++ ExpenseTrackerJS/public/app/dashboard/edit/edit.service.js
@@ -3,10 +3,41 @@
 (function (angular) {
   'use strict';
 
-  const editService = function (ms, elws, acs, aj, us, V) {
+  const editService = function (ms, elws, acs, bs, aj, us, V) {
     const data = {
       expense: {},
+      toRefresh: {
+        accts: {},
+        bills: {}
+      },
       loading: false
+    };
+
+    const initRefresh = function () {
+      data.toRefresh.accts = {};
+      data.toRefresh.bills = {};
+    };
+
+    // store the account ids to be refreshed after modify/delete.
+    const loadRefresh = function () {
+      data.toRefresh.accts[data.expense.accounts.from.id] = data.expense.accounts.from.id;
+      data.toRefresh.accts[data.expense.accounts.to.id] = data.expense.accounts.to.id;
+      if(data.expense.bill && data.expense.bill.id) {
+        data.toRefresh.bills[data.expense.bill.id] = data.expense.bill.id;
+      }
+    };
+
+    // refresh all impacted accounts & bills after modify/delete.
+    const refreshAll = function () {
+      angular.forEach(data.toRefresh.accts, function (value, id) {
+        if(id) {
+          acs.refreshAccount(id);
+        }
+      });
+      angular.forEach(data.toRefresh.bills, function (value, id) {
+        bs.refreshBill(id);
+      });
+      initRefresh();
     };
 
 		// load Bills
@@ -14,20 +45,21 @@
       V.data.bills = dt.data;
     };
     const loadBills = function () {
-      const input = {
-        acctId: data.expense.accounts.from.id,
-      };
+      const input = {acctId: data.expense.accounts.from.id};
 
       aj.query('/dashboard/bills', input, loadBillData);
     };
 
 		// load Page Data
     const loadData = function (dt) {
+      initRefresh();
       data.expense = dt;
       // refresh the 'from' account from TA so what it will have 'billed' attribute.
       if(data.expense.accounts.from.id) {
         data.expense.accounts.from = us.getObjectOf(V.data.accounts, data.expense.accounts.from.id);
       }
+      // store the account ids, bill id to be refreshed after modify/delete.
+      loadRefresh();
 			// initialize Bills TA.
       if (data.expense.accounts.from.billed) {
         loadBills();
@@ -35,38 +67,30 @@
     };
 
 		// modify Expense
-    const loadModifyData = function () {
+    const loadModifyData = function (dt) {
       data.loading = false;
-      elws.modifyItem(data.expense.id);
-
-      if (data.expense.accounts.from.id) {
-        acs.refreshAccount(data.expense.accounts.from.id);
+      us.showMsg('Modify Expense', dt.code);
+      if(dt.code === 0) {
+        elws.modifyItem(data.expense.id);
+        refreshAll();
       }
-      if (data.expense.accounts.to.id) {
-        acs.refreshAccount(data.expense.accounts.to.id);
-      }
-
-      us.showMsg('Modify Expense', 'success');
       $('#model_Modify').modal('hide');
     };
     const modifyExpense = function () {
+      // re-store the account ids, bill id (if these are modified) to be refreshed after save.
+      loadRefresh();
       aj.post('/edit/modify', data.expense, loadModifyData);
       data.loading = true;
     };
 
 		// delete Expense
-    const loadDeleteData = function () {
+    const loadDeleteData = function (dt) {
       data.loading = false;
-      elws.deleteItem(data.expense.id);
-
-      if (data.expense.accounts.from.id) {
-        acs.refreshAccount(data.expense.accounts.from.id);
+      us.showMsg('Delete Expense', dt.code);
+      if(dt.code === 0) {
+        elws.deleteItem(data.expense.id);
+        refreshAll();
       }
-      if (data.expense.accounts.to.id) {
-        acs.refreshAccount(data.expense.accounts.to.id);
-      }
-
-      us.showMsg('Delete Expense', 'success');
       $('#model_Delete').modal('hide');
     };
     const deleteExpense = function () {
@@ -85,6 +109,6 @@
   };
 
   angular.module('dashboard.edit').factory('editService', editService);
-  editService.$inject = ['etmenuService', 'explistwrapperService', 'accountsService',
+  editService.$inject = ['etmenuService', 'explistwrapperService', 'accountsService', 'billsService',
     'ajaxService', 'utilsService', 'VALUES'];
 })(window.angular);
